Remove check for ruby version
The check is _supposed_ to check that the Ruby version is 2.3.3
or greater. However, this is not working as it is supposed to.
I am using Ruby 2.4.4 (distributed with cygwin) and it is giving
an error about it (even though my version is clearly greater
than the specified version). The easiest solution, just delete
the check.